Added get_arg_of_nearest function to utils
